Updated Gradle & AGP
Updated Gradle and AGP to work with JDK 21.0.4 and Android Studio Ladybug.

Warning : This creates a conflict with the upstream (FtcRobotController) but as far as I know there should't be any issues.

If issues do arise, then just revert this commit and use JDK 17 with Gradle when building etc.

If the upstream is updated, merging with this branch should be fine, and the upstream should take precedence. FTC SDK updates should only affect master/FtcRobotController.